renamed the header file for TemplatedSecondaryVertex so it matches the class name
